Some fixes based on feedback from Autodesk. Most are related to warnings.
For mesh writing don't try to be tricky, just get the topology every time and
rely on Alembic's data sharing.

A test with a large model with lots of meshes over a relatively long frame range
(the scene also evaluates quickly), it only adds 15% to the write time.

Try to support reading per vertex UV reading, this needs testing.