From Nicholas Yue: (branch yuenicholas-proceduralinsight)
Mostly CMake changes so it can be built without the boostrap.

Added a bunch of install lines to some of the CMakeLists.txt

For Maya CMake, add devkit/include to include paths. (to accomodate Mac)

Only include SimpleAbcViewer if boost stuff was found.

Don't use tr1 stuff if using MS Visual Studio prior to 2010.

Add a CMake flag for PyILMBase so it can be turned off (for Windows).
Define HALF_EXPORTS for Win32