Revamp AbcStitcher to better handle cases where certain shapes or properties may not exist in some of the files.
We will loosen the requirement of the hierarchy being the same across all
samples, or that there shouldn't be any gaps in time between archives.

Instead default values (or the previous value) will be set for scalar properties
for those missing samples, and for array properties, empty arrays will be set.
(or the previous sample in certain special cases)