Fix export/import when using multi-byte file names
CL 872750, 874437

On Linux/MacOS, the default codepage is UTF8 so asUTF8() happens to
work. But Maya is using the current codepage on Windows. This can be
shift-jis or gbk.  asChar() will return the multi-byte sequence in the
current codepage.  Both Ogawa and HDF5 accepts a std::string as their
parameter. They are not treating the file name parameter as UTF8 as
well. They pass the char* to fopen() or std::ifstream() and both are
using the current codepage.