Feed referenced meshes with an intermediate mesh instead of setting directly
CL 970895, 971367

In Merge mode, Alembic will disconnect the inMesh plug and set the data
via MFnMesh interface directly. This works fine if the mesh is
read/writeable. This approach doesn't work properly when the Maya mesh
is from referenced from a file. Maya doesn't not support deleting
history on a mesh which is referenced..

The solution is to create an intermediate mesh and set data to the
intermediate mesh. Then disconnect the inMesh of the referenced mesh and
connect the intermediate mesh instead. This should generate two
reference edits: disconnectAttr and connectAttr.

Note that Alembic has the mesh data and merge mode should update the
mesh with the data from the Alembic file. So we can't retain the
connection. We need to feed the mesh with the Alembic mesh.