Fix issue where NURBS was not using the previous sample for the optional position weights and velocities when those properties were set at some point but the current sample did not define them.  (rank 0 dimensions)
Add some tests for position weights, and add peculiar sampling of empty and non
empty array property samples.

Be a little more lenient when all weights are near 1.0 in Maya.

If a rank 0 array sample reaches AbcCoreOgawa treat it as a rank 1 of size 0.
AbcCoreHDF5 will throw on this, but we can be a little more lenient going
forward.