WIP.
Introduce SparseFlag to Arguments that will eventually suppress metadata
for schemas.  Should this even be here, or should it be an optional bool on the
schema constructor itself....

Also, while we were looking around, address a long outstanding annoyance with
the templated constructors on the write side in the Abc layer.

We also deprecate the dumb WrapExistingFlag. (and we'll do TopFlag for
OObject next)

Need to do the same thing on the read side, and we should consider whether
we really want to update some of the constructors to accept
AbcA::CompoundWriterPtr and OCompoundProperty, or if we should only have 1.

Also some light tab to spaces touchup to OPolyMesh.