Prune support
Support for setting the "prune" flag in objects and properties. The
prune flag is set on an object or property that exists in file1's
hierarchy but is missing from file2's hierarchy.

This is useful for using AbcDiff to generate layer files that can
describe the removal of an object or property. Corresponding API code in
AbcLayer will notice the prune flag and remove references of the
object/property from the base Alembic file, accordingly.