Ability to connect the AlembicNode to first shape of the current parent
It's a a simple option which determines if it's necessary to use the first
available DAG node of the current parent node when it's not possible to find the
DAG node to connect it to the AlembicNode. It can be useful when the Alembic
file was exported after Maya renamed the shapes nodes of an object (it happens
when deforming reference objects). Usually, in Maya, a transform node contains
only one shape and it works perfectly.