Fix per-face shader assignments are not merged if the mesh to merge is an intermediate object.
MAYA-74786
If we add a deformer to a mesh, the mesh is pushed into the construction history of
the deformed mesh. The AbcImport merges Alembic to the original mesh by name matching.
This is OK since we keep track of the original mesh and update it only. But we should
update the per-face shader assignment of the deformed mesh. Because the deformed mesh
is the mesh for shading. It's meaningless to assign the faces of an intermediate object
to a shading group.