Alembic file translator should resolve in the same way as AbcImport command
MAYA-74119

Alembic should resolve file path by using kInputFile for relative and
dirmapping. This was done for AbcImport command. The file translator
should do the same. This affects File -> Import and File -> Reference.

The issue is Maya can't resolve C:/path/ball.abc as an absolute path
in kNone mode. On Linux, C:/path/ball.abc is a relative path. Python
os.path.isabs() is the same. So we have to treat C:/path/ball.abc
as a relative path and prepend the current workspace path to it:
/path/to/workspace/C:/path/ball.abc. This is a bad path.. The
behavior is not wrong. We can't say C:/path/ball.abc is an absolute
path on Linux..