Change up the throw message for bad meta data to make it easier to test. Fuzzer Issue 25081 correctly throws.
Fuzzer Issues 25166, 25175, 25185, 25204, and 25236 all seem to be different
variations on an illegal meta data index value coming through.  We more
carefully check for that now.

Fuzzer Issue 25192 turned up a leak in OrData where the raw pointer allocated
for the children was not being released because the constructor threw.
Putting it into a unique_ptr fixes this problem.