WIP prep for Imath 3 by looking for the Imath module.
We need to find the correct way to find if PyImath exists
and include/link to PyImath from Imath when using Imath 3.

We weren't looking for OpenEXR so remove the cmake for it.
Tweak AlembicIlmBase and AlembicPyIlmBase to look for the Imath package
(which will come from Imath3) and then fall back on the old way for finding
IlmBase.

We also still need to figure out how to run our python test suite
automatically from CMake.