Address Fuzzer Issue 33685.
While trying to test the reported fuzzer test, it was noticed that if the
compound properties were NOT walked (and only the object hierarchy) it would
recurse forever.

To address this we reworked IGroup::getGroup to make sure the group wasnt
obviously bad.  We also got rid of the throw if we notice the child and parent
point at each other, and instead return an empty shared_ptr.

Doing this caused us to rework the reported exception in a few of our fuzzer
tests.

We also test walking JUST the object hierachy, before walking both the object
hiearchy and the property hierarchy where appropriate.