Fixes to make Alembic exported config work for OpenEXR 2.x or Imath 3.x
I believe the way it currently was, the exported AlembicConfig.cmake
was fine for if Imath3 is being used, but for building against older
OpenEXR 2.x, would not work properly.

The change: AlembicIlmBase.cmake sets a variable that reveals whether
Imath3 was found or if we fall back to useing OpenEXR 2.x. Then we use
that to alter what the exported Alembic.config.cmake ought to say in
terms of whether to find_dependency(Imath) or find_dependency(IlmBase).

This extra smidge of complexity can be removed when Imath 3.x is the
true minimum. But given that OpenEXR 2.4 is (unwisely, IMHO) the
choice for VFX Platform 2021 and in the current set of DCC versions, I
believe we need to support OpenEXR 2.4 as a first class citizen for
just a little bit longer.

Signed-off-by: Larry Gritz <lg@larrygritz.com>