Add a couple more tests to the fuzzer and walk the hierarchy catching exceptions so we can look for potentially illegal accesses for a couple of files.
We may try to expand this to all of them in a future commit.

Update a deprecated CMake call.

The new tests turned up a buffer overflow because of the potential difference
in our read dimensions and the size of our IDataPtr.

For now we make sure not to read past what we allocated, but in the future we
may throw if our dimensions are mismatched with the size of our data. (TBD)